refactor(cf): move upcoming contests to codeforces.go
Move logic about updating list of upcoming contests from listContests.go
to codeforces.go. This better reflects where these functions are used,
because they will also be used in other codeforces commands than !cf
contests.
